* Edited the Core Data format (made shields a float, used to be an int). Be sure to delete previous app installs to avoid save crashes. * Shields is now implemented. You get them at level 10. Reminder use the hero editor in the dev toolbox to change your level. * To put shields up drag from one blue orb to the other. Shields will deteriorate over time. It currently lasts 10 seconds. While shields are up incoming damage is halved (I could edit the dmg reduction to be in ratio to the shield power, thoughts?). * The camera background of the battle view has been taken out (per last meeting discussions).

+++ Game Engine/RQMob.h
@@ -32,7 +32,7 @@ typedef NSUInteger RQElementalType;
 @property (readwrite, assign) NSInteger experiencePoints;
 @property (readwrite, assign) float stamina;
 @property (readwrite, assign) float staminaRegenRate;
-@property (readwrite, assign) NSInteger secondsLeftOfShields;
+@property (readwrite, assign) float secondsLeftOfShields;
 
 - (NSInteger)randomAttackValueAgainstMob:(RQMob *)mob withWeaponOfType:(RQElementalType)weaponType;
